requirements.txt
mediapipe>=0.8.6.2
numpy>=1.21.5
opencv_python>=4.5.1.48
pandas>=1.3.5
Pillow>=9.2.0
py>=1.10.0
pynput>=1.7.6
PyQt5>=5.15.7
scikit_learn>=1.1.1
tqdm>=4.62.3
